se completo la integración del sistema de rutina, solo falta arreglar el pointer para que cierre al agregar tarea

+++ app/src/main/java/com/example/equipohawknetwork/routines/RoutinesActivity.kt
@@ -68,7 +68,7 @@ class RoutinesActivity : AppCompatActivity() {
                     muscleGroup = group,
                     sets = sets,
                     reps = reps,
-                    weightKg = weight,
+                    weightKg = weight.toDouble(), // 👈 Float -> Double
                     restSec = rest
                 )
             )
@@ -86,7 +86,7 @@ class RoutinesActivity : AppCompatActivity() {
                     muscleGroup = group,
                     sets = sets,
                     reps = reps,
-                    weightKg = weight,
+                    weightKg = weight.toDouble(), // 👈 Float -> Double
                     restSec = rest
                 )
             )
